Refactor Per-App Volume logic into a standalone module
- Extract Per-App Volume implementation from `VolumePanel` into a new `PerAppVolume` module.
- Introduce `ExtendedSlider` view with Material3 and Dynamic Colors support.
- Refactor `ExtendedFAB` to use a shared `createMonetM3Context` utility for consistent styling.
- Register `PerAppVolume` module in `EntryList`.
- Update volume panel string resources and descriptions for better clarity.
- Enhance the Per-App Volume UI by replacing standard `SeekBar` with `ExtendedSlider`.
