+++ app/src/main/java/com/drdisagree/iconify/xposed/EntryList.kt
@@ -41,6 +41,7 @@ import com.drdisagree.iconify.xposed.modules.statusbar.StatusbarLogo
 import com.drdisagree.iconify.xposed.modules.statusbar.StatusbarMisc
 import com.drdisagree.iconify.xposed.modules.statusbar.SwapSignalNetworkType
 import com.drdisagree.iconify.xposed.modules.statusbar.SwapWiFiCellular
+import com.drdisagree.iconify.xposed.modules.volume.PerAppVolume
 import com.drdisagree.iconify.xposed.modules.volume.VolumePanel
 import com.drdisagree.iconify.xposed.utils.HookCheck
 
@@ -85,6 +86,7 @@ object EntryList {
         DualStatusbar::class.java,
         StatusbarMisc::class.java,
         VolumePanel::class.java,
+        PerAppVolume::class.java,
         ColorizeNotificationView::class.java,
         AppIconInNotification::class.java,
         HeadsUpBlur::class.java,
